feat: custom auth controllers
