+++ book_keeper/lib/auth/presentation/pages/login_screen.dart
@@ -1,4 +1,4 @@
-import 'package:book_keeper/auth/presentation/controllers/login_page_controller.dart';
+import 'package:book_keeper/auth/presentation/controllers/google_auth_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -12,8 +12,8 @@ class LoginScreen extends ConsumerStatefulWidget {
 class _LoginScreenState extends ConsumerState<LoginScreen> {
   @override
   Widget build(BuildContext context) {
-    final state = ref.watch(loginPageControllerProvider);
-    ref.listen(loginPageControllerProvider, (_, state) {
+    final state = ref.watch(googleAuthControllerProvider);
+    ref.listen(googleAuthControllerProvider, (_, state) {
       if (!state.isLoading && state.hasError) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text(state.error.toString())),
@@ -30,7 +30,7 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
         child: state.isLoading
             ? const CircularProgressIndicator()
             : ElevatedButton(
-                onPressed: () => ref.read(loginPageControllerProvider.notifier).signInWithGoogle(),
+                onPressed: () => ref.read(googleAuthControllerProvider.notifier).signInWithGoogle(),
                 child: const Text('Sign in with Google'),
               ),
       ),
